scripts/requirements.txt
python-docx>=1.1.0
msal>=1.28.0
requests>=2.32.0
